Backend-main/massuh-datalake-main/apresentame-webhook/requirements.txt
functions-framework==3.*
pandas
google-cloud-bigquery
google-auth
pandas-gbq